# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Phase 1 control-plane and gateway contract for Command Post.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [overview]
      summary: List recent operational events
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Recent events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OperationalEvent"
  /providers:
    get:
      operationId: listProviders
      tags: [providers]
      summary: List configured providers
      responses:
        "200":
          description: Providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Provider"
    post:
      operationId: createProvider
      tags: [providers]
      summary: Register a provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderInput"
      responses:
        "201":
          description: Registered provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
        "400":
          $ref: "#/components/responses/BadRequest"
  /providers/{id}/status:
    patch:
      operationId: updateProviderStatus
      tags: [providers]
      summary: Enable or disable a provider
      parameters:
        - $ref: "#/components/parameters/ProviderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderStatusInput"
      responses:
        "200":
          description: Updated provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /routing/policy:
    get:
      operationId: getRoutingPolicy
      tags: [routing]
      summary: Get routing policy
      responses:
        "200":
          description: Routing policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutingPolicy"
    put:
      operationId: updateRoutingPolicy
      tags: [routing]
      summary: Update routing policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoutingPolicyInput"
      responses:
        "200":
          description: Updated routing policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutingPolicy"
...
  /security/events:
    get:
      operationId: listSecurityEvents
      tags: [security]
      summary: List recent security events
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Security events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SecurityEvent"
...
components:
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    ProviderId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    OperationalEvent:
      type: object
      required: [id, level, title, detail, occurredAt]
      properties:
        id:
          type: string
        level:
          type: string
          enum: [info, warning, error, critical]
        title:
          type: string
        detail:
          type: string
        occurredAt:
          type: string
          format: date-time
    Provider:
      type: object
      required: [id, name, kind, status, health, latency, errorRate, models, lastCheckedAt]
      properties:
        id:
          type: string
        name:
          type: string
        kind:
          type: string
          enum: [hosted, local]
        status:
          type: string
          enum: [enabled, disabled]
        health:
          type: string
          enum: [healthy, degraded, unavailable]
        latency:
          type: number
        errorRate:
          type: number
        models:
          type: integer
        lastCheckedAt:
          type: string
          format: date-time
    ProviderInput:
      type: object
      required: [name, kind]
      properties:
        name:
          type: string
          minLength: 1
        kind:
          type: string
          enum: [hosted, local]
    ProviderStatusInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [enabled, disabled]
...
    RoutingPolicy:
      type: object
      required: [strategy, maxAttempts, costGuardrail, latencyTargets, updatedAt]
      properties:
        strategy:
          type: string
          enum: [lowest_predicted_latency]
        maxAttempts:
          type: integer
        costGuardrail:
          type: number
        latencyTargets:
          type: object
          required: [standard, longContext, structured]
          properties:
            standard:
              type: integer
            longContext:
              type: integer
            structured:
              type: integer
        updatedAt:
          type: string
          format: date-time
    RoutingPolicyInput:
      type: object
      required: [costGuardrail, standardLatency, longContextLatency, structuredLatency]
      properties:
        costGuardrail:
          type: number
          minimum: 0
          maximum: 100
        standardLatency:
          type: integer
          minimum: 1
        longContextLatency:
          type: integer
          minimum: 1
        structuredLatency:
          type: integer
          minimum: 1
...
    SecurityEvent:
      type: object
      required: [id, severity, title, detail, occurredAt, status]
      properties:
        id:
          type: string
        severity:
          type: string
          enum: [info, warning, error, critical]
        title:
          type: string
        detail:
          type: string
        occurredAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [open, acknowledged, resolved]
...